% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeEnrichment.R
\name{CellTypeEnrichment}
\alias{CellTypeEnrichment}
\title{Cell Type Enrichment Analysis for Gene Expression scRNA-seq data}
\usage{
CellTypeEnrichment(
  SampleName,
  SamplesMetadata,
  ReadCount,
  GTF,
  BackendDataDir,
  tSNECellranger,
  UMAPCellranger,
  RawFeatureDir,
  FilteredFeatureBarcodes
)
}
\arguments{
\item{SampleName}{Name of an indicated sample}

\item{SamplesMetadata}{Metadata of samples including the following headers: Project Number,    LIMS ID,    Sample Name,    Index    Library Type,    Genome,    Flowcell ID,    Lane Number,    Sequencing ID}

\item{ReadCount}{Gene-cell count matrix from 10X CellRanger count}

\item{GTF}{We convert Ensembl ids to gene names/symbols by parsing this GTF (General Transfer Format) file}

\item{BackendDataDir}{We used Human Protein Atlas database (version 21.0) to generate a repository of reference gene sets that are exclusively expressed in each cell type}

\item{tSNECellranger}{tSNE projections from 10X CellRanger count}

\item{UMAPCellranger}{UMAP projections from 10X CellRanger count}

\item{RawFeatureDir}{A folder including raw feature-barcode matrices from 10X CellRanger count (barcodes.tsv.gz, features.tsv.gz, matrix.mtx.gz)}

\item{FilteredFeatureBarcodes}{#'A folder including filtered feature-barcode matrices from 10X CellRanger count (barcodes.tsv.gz)}
}
\value{
You can find the results in R object under title of 'RESULTsGenomicFeatures' and 'RESULTsChromatinState'.
}
\description{
Cell type annotation on scRNA-Seq data is a pre-step for generating an interactive QC report with scQCEA. scQCEA provides a function called CellTypeEnrichment, for automatic cell type identification and visualization on the gene-by-cell count matrix.
}
\examples{
library("scQCEA")
csQCEAdir <- system.file("extdata", package = "scQCEA")
DataTyep <- '10X-gex'
SampleName <- '481207_03'
SamplesMetadata = paste(csQCEAdir, 'Inputs/samples.metadata', sep = '/' )
ReadCount = paste(csQCEAdir, 'Inputs', DataTyep, SampleName, 'outs', sep = '/')
GTF = paste(csQCEAdir, 'ensembl_human.txt', sep = '/')
BackendDataDir = paste(csQCEAdir, 'ReferenceGeneSets/', sep = '/')
tSNECellranger = paste(csQCEAdir, 'Inputs', DataTyep, SampleName, '/outs/analysis/tsne/gene_expression_2_components', sep = '/')
UMAPCellranger =  paste(csQCEAdir, 'Inputs', DataTyep, SampleName, '/outs/analysis/umap/gene_expression_2_components', sep = '/')
RawFeatureDir = paste(csQCEAdir, 'Inputs', DataTyep, SampleName, 'outs/raw_feature_bc_matrix', sep = '/')
FilteredFeatureBarcodes = paste(csQCEAdir, 'Inputs', DataTyep, SampleName, 'outs/filtered_feature_bc_matrix', sep = '/')
CellTypeEnrichment(SampleName, SamplesMetadata, ReadCount, GTF, BackendDataDir, tSNECellranger, UMAPCellranger, RawFeatureDir, FilteredFeatureBarcodes ) 
}
\author{
{Isar Nassiri}
}
