% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInteractiveQCReport.R
\name{GenerateInteractiveQCReport}
\alias{GenerateInteractiveQCReport}
\title{Generate Interactive QC Report}
\usage{
GenerateInteractiveQCReport()
}
\value{
You can find the results in R object under title of 'RESULTsGenomicFeatures' and 'RESULTsChromatinState'.
}
\description{
An interactive QC report automatically will be generated in one HTML file, including four sections: experimental workflow, data processing workflow, sample information and QC metrics, data analysis and quality control.
}
\examples{
library("scQCEA")
GenerateInteractiveQCReport()
}
\author{
{Isar Nassiri}
}
